requirements.txt
Flask
opencv-python
numpy
mediapipe
pillow
